# Pre-allocate error vector in TRY